Workaround for BOT_EQUIV.
The problem is that now two calls may be bound to the same variable, so one of them needs to be translated into a _reif directly. Ideally, a vardecl should be able to have a list of calls that is equivalent to that variable, to keep the tree structure intact.