Store if variable has a domain computed from its right hand side
This is used in the optimisation phase: a variable that is not used can be removed even if it has a domain (which corresponds to a constraint using that variable), IF the variable is "implied free", i.e. the domain is a superset of the right hand side of the variable.  This is approximated by storing if we have computed the bounds of the variable from the RHS.