Create output model
The output model is currently a direct copy of the original, removing all unnecessary items and all domain constraints. Some things are currently not supported, such as tests, functions or predicates in the output item. We will also need to reconsider output once we have more interesting optimisations that remove variables (e.g. when replacing a var int with an array[int] of var bool).