Workaround for problem with reified predicates that have undefined arguments.
If the flattening of a reified predicate creates a subexpression that has previously been flattened in a boolean context and that had non-trivial partiality, the expression would have been taken from the CSE map and constrained to be total (in the root context). This is incorrect. The workaround is to ignore all partiality coming from CSE expressions when flattening _reif predicates. The upshot is that _reif predicates cannot introduce new partiality.