Fix for path handling when 'needRangeDomains' is active.
'needRangeDomains' causes explicit set_in constraints to be added to
the FlatZinc. The CallStackItem that was being constructed for these
was not unique, including only the second argument of the set_in
call.

The fix involves extracting the existing path annotation for the
variable and using this as the CallStackItem. This should allow
the paths to be matched across compilations (for multi-pass)