Change ASTString to use String Interning
This changes every ASTString to exists at most once. When a new
ASTString is about to be allocated, a hash map is consulted to reuse the
ASTString if it already exists. This safes memory and allows for ==
comparisons in O(1).

This also changes removes the `str()` method. This ensures that the
string is not implicitly copied. A std::string copy can still be made
using the `c_str()` method as the argument to the std::string constructor