Handle mzn_check_var and mzn_check_enum_var after collecting VarDecls for the output model
There are some cases where variable declarations are copied to the output before that id is visited in CollectVarDecls. This prevents `mzn_check_var` and `mzn_check_enum_var` annotations from being added to the copies. Instead just do this at the end, so we can be sure they're processed.