Handle more binary operator expressions as linear expressions
When determining whether a binary operator can be turned into a linear constraint, if one of the arguments is an identifier that's bound to a non-linear expression, still generate a linear constraint using that identifier. Previously, an equality constraint was generated instead (which could often not be optimised away), rather than e.g. simply limiting the domain of the variable of the identifier.