Make the comparator CmpExpIdx respect the std::strict_weak_ordering requirements.
With the current code, if there are two id Expressions: `id1` and `id2` and one non-id expression `expr`, we could have `id1 < id2` and `id2 < expr < id1`. This CL makes ids to be ordered before non-ids.

Also remove call to Expression::equal(), which also made a checker complain the ordering was not strict, probably because it is not transitive.